Bug 1377272 - Making ParserContext less intertwined with Parser;r=shu+381259
To implement the BinJS decoder of Bug 1377007, we need to access ParserContext without a Parser. This patch makes it possible to construct one.

MozReview-Commit-ID: Fx3S0UkU7Hq

--HG--
extra : rebase_source : 7470b03fb45bd8d80ffd66b1507d1491b5e3158f
